refactor(multipage): Rearrange streamlit code into multiple pages
